+++ GUI/LightingGUI/customarraypage.cpp
@@ -4,15 +4,15 @@
  *
  */
 
-#include "ArrayColorsPage.h"
-#include "ui_ArrayColorsPage.h"
+#include "CustomArrayPage.h"
+#include "ui_CustomArrayPage.h"
 #include "icondata.h"
 
 #include <QSignalMapper>
 
-ArrayColorsPage::ArrayColorsPage(QWidget *parent) :
+CustomArrayPage::CustomArrayPage(QWidget *parent) :
     QWidget(parent),
-    ui(new Ui::ArrayColorsPage) {
+    ui(new Ui::CustomArrayPage) {
     ui->setupUi(this);
 
     // --------------
@@ -21,22 +21,23 @@ ArrayColorsPage::ArrayColorsPage(QWidget *parent) :
 
     const int COLOR_ARRAY_PRESET = 10;
     connect(ui->arraySlider, SIGNAL(valueChanged(int)), this, SLOT(colorsUsedChanged(int)));
-    ui->arraySlider->setSliderColorBackground({0,255,0});
+    ui->arraySlider->setSliderColorBackground(QColor(0,255,0));
     ui->arraySlider->slider->setMaximum(COLOR_ARRAY_PRESET * 10);
     ui->arraySlider->slider->setTickInterval(10);
     ui->arraySlider->slider->setTickPosition(QSlider::TicksBelow);
     ui->arraySlider->setSnapToNearestTick(true);
     ui->arraySlider->setMinimumPossible(true, 20);
 
+
     // --------------
     // Setup Color Array Array (array array  ... array array...)
     // --------------
 
-    mArrayColorsButtons = std::shared_ptr<std::vector<QPushButton*> >(new std::vector<QPushButton*>(COLOR_ARRAY_PRESET, nullptr));
+    mArrayColorsButtons = std::shared_ptr<std::vector<QToolButton*> >(new std::vector<QToolButton*>(COLOR_ARRAY_PRESET, nullptr));
     mArrayColorsIconData = std::shared_ptr<std::vector<IconData> >(new std::vector<IconData>(COLOR_ARRAY_PRESET));
     QSignalMapper *arrayButtonsMapper = new QSignalMapper(this);
     for (int i = 0; i < COLOR_ARRAY_PRESET; ++i) {
-        (*mArrayColorsButtons.get())[i] = new QPushButton;
+        (*mArrayColorsButtons.get())[i] = new QToolButton;
         (*mArrayColorsIconData.get())[i] = IconData(80,80);
         (*mArrayColorsButtons.get())[i]->setIcon((*mArrayColorsIconData.get())[i].renderAsQPixmap());
         ui->arrayColorsLayout->addWidget((*mArrayColorsButtons.get())[i], 0 , i);
@@ -69,54 +70,53 @@ ArrayColorsPage::ArrayColorsPage(QWidget *parent) :
     (*mPageButtons.get())[5] = ui->barsMovingButton;
 
     // loop through the buttons and set up their clicks to a mapper
-    QSignalMapper* signalMapper = new QSignalMapper(this);
-    for (uint i = 0; i < mPageButtons->size(); i++)
-    {
+    QSignalMapper *signalMapper = new QSignalMapper(this);
+    for (uint i = 0; i < mPageButtons->size(); i++) {
         QToolButton* button =(*mPageButtons.get())[i];
         button->setCheckable(true);
         connect(button, SIGNAL(clicked(bool)), signalMapper, SLOT(map()));
     }
 
-    signalMapper->setMapping(ui->glimmerButton, (int)ELightingMode::eLightingModeArrayGlimmer);
-    signalMapper->setMapping(ui->randomIndividualButton, (int)ELightingMode::eLightingModeArrayRandomIndividual);
-    signalMapper->setMapping(ui->randomSolidButton, (int)ELightingMode::eLightingModeArrayRandomSolid);
-    signalMapper->setMapping(ui->fadeButton, (int)ELightingMode::eLightingModeArrayFade);
-    signalMapper->setMapping(ui->barsSolidButton, (int)ELightingMode::eLightingModeArrayBarsSolid);
-    signalMapper->setMapping(ui->barsMovingButton, (int)ELightingMode::eLightingModeArrayBarsMoving);
+    signalMapper->setMapping(ui->glimmerButton, (int)ELightingMode::eMultiGlimmer);
+    signalMapper->setMapping(ui->randomIndividualButton, (int)ELightingMode::eMultiRandomIndividual);
+    signalMapper->setMapping(ui->randomSolidButton, (int)ELightingMode::eMultiRandomSolid);
+    signalMapper->setMapping(ui->fadeButton, (int)ELightingMode::eMultiFade);
+    signalMapper->setMapping(ui->barsSolidButton, (int)ELightingMode::eMultiBarsSolid);
+    signalMapper->setMapping(ui->barsMovingButton, (int)ELightingMode::eMultiBarsMoving);
 
     connect(signalMapper, SIGNAL(mapped(int)), this, SLOT(modeChanged(int)));
 }
 
-ArrayColorsPage::~ArrayColorsPage() {
+CustomArrayPage::~CustomArrayPage() {
     delete ui;
 }
 
 
-void ArrayColorsPage::highlightModeButton(ELightingMode lightingMode) {
+void CustomArrayPage::highlightModeButton(ELightingMode lightingMode) {
 
     for (uint i = 0; i < mPageButtons->size(); i++) {
         QToolButton* button =(*mPageButtons.get())[i];
         button->setChecked(false);
     }
 
-    if (lightingMode == ELightingMode::eLightingModeArrayGlimmer) {
+    if (lightingMode == ELightingMode::eMultiGlimmer) {
         ui->glimmerButton->setChecked(true);
-    } else if (lightingMode == ELightingMode::eLightingModeArrayRandomIndividual) {
+    } else if (lightingMode == ELightingMode::eMultiRandomIndividual) {
         ui->randomIndividualButton->setChecked(true);
-    } else if (lightingMode == ELightingMode::eLightingModeArrayRandomSolid) {
+    } else if (lightingMode == ELightingMode::eMultiRandomSolid) {
         ui->randomSolidButton->setChecked(true);
-    } else if (lightingMode == ELightingMode::eLightingModeArrayFade) {
+    } else if (lightingMode == ELightingMode::eMultiFade) {
         ui->fadeButton->setChecked(true);
-    } else if (lightingMode == ELightingMode::eLightingModeArrayBarsSolid) {
+    } else if (lightingMode == ELightingMode::eMultiBarsSolid) {
         ui->barsSolidButton->setChecked(true);
-    } else if (lightingMode == ELightingMode::eLightingModeArrayBarsMoving) {
+    } else if (lightingMode == ELightingMode::eMultiBarsMoving) {
         ui->barsMovingButton->setChecked(true);
     }
 }
 
-void ArrayColorsPage::selectArrayColor(int index) {
+void CustomArrayPage::selectArrayColor(int index) {
     mCurrentColorPickerIndex = index;
-    ui->colorPicker->chooseColor(mData->colors[mCurrentColorPickerIndex], false);
+    ui->colorPicker->chooseColor(mData->customArray()[mCurrentColorPickerIndex], false);
     updateColorArray();
     for (int i = 0; i < mData->colorCount(); ++i) {
         (*mArrayColorsButtons.get())[i]->setChecked(false);
@@ -127,7 +127,7 @@ void ArrayColorsPage::selectArrayColor(int index) {
     (*mArrayColorsButtons.get())[mCurrentColorPickerIndex]->setStyleSheet("border: 2px solid white");
 }
 
-void ArrayColorsPage::updateColorArray() {
+void CustomArrayPage::updateColorArray() {
     for (int i = 0; i < mData->colorCount(); ++i) {
         (*mArrayColorsButtons.get())[i]->setEnabled(true);
         (*mArrayColorsButtons.get())[i]->setIcon((*mArrayColorsIconData.get())[i].renderAsQPixmap());
@@ -145,36 +145,27 @@ void ArrayColorsPage::updateColorArray() {
 // Slots
 // ----------------------------
 
-void ArrayColorsPage::colorsUsedChanged(int newColorsUsed) {
+void CustomArrayPage::colorsUsedChanged(int newColorsUsed) {
     int totalColorsUsed = newColorsUsed / 10;
     ELightingMode mode = mData->currentMode();
     if (totalColorsUsed != mData->colorsUsed()) {
         mData->colorsUsed(totalColorsUsed);
         updateColorArray();
         updateIcons();
-        if (mode == ELightingMode::eLightingModeArrayBarsMoving
-            || mode == ELightingMode::eLightingModeArrayBarsSolid
-            || mode == ELightingMode::eLightingModeArrayFade
-            || mode == ELightingMode::eLightingModeArrayGlimmer
-            || mode == ELightingMode::eLightingModeArrayRandomIndividual
-            || mode == ELightingMode::eLightingModeArrayRandomSolid) {
-            mData->currentMode(mode);
-            mComm->sendArrayModeChange(mode, totalColorsUsed);
-        }
-
+        mComm->sendCustomArrayCount(mData->colorsUsed());
     }
 }
 
 
-void ArrayColorsPage::modeChanged(int newMode) {
+void CustomArrayPage::modeChanged(int newMode) {
     mData->currentMode((ELightingMode)newMode);
-    mComm->sendArrayModeChange(mData->currentMode(), mData->colorsUsed());
+    mComm->sendArrayModeChange(mData->currentMode());
     highlightModeButton(mData->currentMode());
     emit updateMainIcons();
 }
 
-void ArrayColorsPage::colorChanged(QColor color) {
-    mData->colors[mCurrentColorPickerIndex] = color;
+void CustomArrayPage::colorChanged(QColor color) {
+    mData->customArray()[mCurrentColorPickerIndex] = color;
 
     mComm->sendArrayColorChange(mCurrentColorPickerIndex, color);
     ui->arraySlider->setSliderColorBackground(mData->colorsAverage());
@@ -188,7 +179,7 @@ void ArrayColorsPage::colorChanged(QColor color) {
 // Protected
 // ----------------------------
 
-void ArrayColorsPage::showEvent(QShowEvent *event) {
+void CustomArrayPage::showEvent(QShowEvent *event) {
   Q_UNUSED(event);
   // update all the mode icons to the current colors.
   updateIcons();
@@ -208,10 +199,10 @@ void ArrayColorsPage::showEvent(QShowEvent *event) {
 // Private
 // ----------------------------
 
-void ArrayColorsPage::updateIcons() {
+void CustomArrayPage::updateIcons() {
 
     for (int i = 0; i < mData->colorsUsed(); ++i) {
-        (*mArrayColorsIconData.get())[i].setSolidColor(mData->colors[i]);
+        (*mArrayColorsIconData.get())[i].setSolidColor(mData->customArray()[i]);
         (*mArrayColorsButtons.get())[i]->setEnabled(true);
         (*mArrayColorsButtons.get())[i]->setIcon((*mArrayColorsIconData.get())[i].renderAsQPixmap());
     }
@@ -224,37 +215,37 @@ void ArrayColorsPage::updateIcons() {
     IconData buttonData = IconData(ui->glimmerButton->iconSize().width(),
                                    ui->glimmerButton->iconSize().height(),
                                    mData);
-    buttonData.setArrayGlimmer();
+    buttonData.setArrayGlimmer(EColorPreset::eCustom);
     ui->glimmerButton->setIcon(buttonData.renderAsQPixmap());
 
     buttonData = IconData(ui->fadeButton->iconSize().width(),
                           ui->fadeButton->iconSize().height(),
                           mData);
-    buttonData.setArrayFade();
+    buttonData.setArrayFade(EColorPreset::eCustom);
     ui->fadeButton->setIcon(buttonData.renderAsQPixmap());
 
     buttonData = IconData(ui->randomSolidButton->iconSize().width(),
                           ui->randomSolidButton->iconSize().height(),
                           mData);
-    buttonData.setArrayRandomSolid();
+    buttonData.setArrayRandomSolid(EColorPreset::eCustom);
     ui->randomSolidButton->setIcon(buttonData.renderAsQPixmap());
 
     buttonData = IconData(ui->randomIndividualButton->iconSize().width(),
                           ui->randomIndividualButton->iconSize().height(),
                           mData);
-    buttonData.setArrayRandomIndividual();
+    buttonData.setArrayRandomIndividual(EColorPreset::eCustom);
     ui->randomIndividualButton->setIcon(buttonData.renderAsQPixmap());
 
     buttonData = IconData(ui->barsSolidButton->iconSize().width(),
                           ui->barsSolidButton->iconSize().height(),
                           mData);
-    buttonData.setArrayBarsSolid();
+    buttonData.setArrayBarsSolid(EColorPreset::eCustom);
     ui->barsSolidButton->setIcon(buttonData.renderAsQPixmap());
 
     buttonData = IconData(ui->barsMovingButton->iconSize().width(),
                           ui->barsMovingButton->iconSize().height(),
                           mData);
-    buttonData.setArrayBarsMoving();
+    buttonData.setArrayBarsMoving(EColorPreset::eCustom);
     ui->barsMovingButton->setIcon(buttonData.renderAsQPixmap());
 }